allow to create optionalRef's
When we have a `ref`, then we want to put it on the element, but if you
are using a `Fragment` then this is not necessary and thus optional.

This allows us to get a ref to an element in order to determine the
`ownerDocument` if a ref is present. If not, then we will fallback to
the normal `document`.